requirements.txt
streamlit
pandas
pickle-mixin
google-auth
google-auth-oauthlib
google-api-python-client
nltk
python-dateutil
scikit-learn